Check zero emission if so set to norm stake
Zero incentive results when all neurons either did not set weights, or set outdated weights, or set self-weights, resulting in zero weights and consequently zero incentive. This could happen at block 0 or when all uids have reregistered without new weights set on these.

In the case of zero weights the emission should be proportional to stake, and if stake is zero then do not mask inactive stake. If unmasked stake is zero then emission will be zero, and this would need to be handled in epoch caller function.