Refactor IsValidator -> ValidatorPermit
To avoid confusion, since a top-stake uid does not necessarily validate, so we grant a validator permit instead with the choice to validate left to the uid. Otherwise, further development may assume is_validator is literally true, although it is just a permission to set weights and have bonds stored, and not necessarily an actual validator.